fix: improve Settings
