+++ packages/core/src/cloud/LlamaCloudRetriever.ts
@@ -3,7 +3,7 @@ import { globalsHelper } from "../GlobalsHelper.js";
 import type { NodeWithScore } from "../Node.js";
 import { ObjectType, jsonToNode } from "../Node.js";
 import type { BaseRetriever, RetrieveParams } from "../Retriever.js";
-import { getCurrentCallbackManager } from "../callbacks/CallbackManager.js";
+import { getCurrentCallbackManager } from "../Settings.js";
 import type { ClientParams, CloudConstructorParams } from "./types.js";
 import { DEFAULT_PROJECT_NAME } from "./types.js";
 import { getClient } from "./utils.js";
